函数打包Lib.java，并实现统计单词数功能
